pyproject.toml
[tool.poetry]
name = "yarawesome"
version = "0.1.0"
description = ""
authors = ["JaminB <jamin.becker@huntresslabs.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
watchdog = "^3.0.0"
python-dotenv = "^1.0.0"
plyara = "^2.1.1"
django-rest-framework = "^0.1.0"
django-mathfilters = "^1.0.0"
patool = "^1.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
